+++ R/GPsInElevGaps.R
@@ -13,10 +13,12 @@
 #' @param chunkContigs TRUE or FALSE, If TRUE and `mode`="metagenome", contigs longer
 #' than the `chunkSize` will be 'chunked' into smaller subsets and pattern-matching
 #' will be performed on each subset. Default is FALSE.
+#' @param chunkSize If `mode`="genome" OR if `mode`="metagenome" and `chunkContigs`=TRUE,
+#' chunk the genome or contigs, respectively, into smaller subsets for pattern-matching.
 #' @importFrom stringr str_detect str_extract_all regex
 #' @importFrom dplyr bind_rows
 #' @keywords internal
-GPsInElevGaps <- function(elevGapSummList, windowSize, gffTSV, mode, chunkContigs) {
+GPsInElevGaps <- function(elevGapSummList, windowSize, gffTSV, mode, chunkContigs, chunkSize) {
   colnames(gffTSV) <- c("seqid", "source", "type", "start", "end", "score", "strand", "phase", "attributes")
   if (TRUE %in% (str_detect(gffTSV[,9], regex('product', ignore_case = T)))){
     product <- str_extract_all(gffTSV [,9], regex("(?<=product=)[\\s\\S]*",ignore_case = T))
@@ -26,13 +28,17 @@ GPsInElevGaps <- function(elevGapSummList, windowSize, gffTSV, mode, chunkContig
     trueRefName <- elevGapSummList[[i]][[8]]
     if(mode == "metagenome"){
       refName <- elevGapSummList[[i]][[8]]
-    if(chunkContigs==TRUE){
+    if(chunkContigs == TRUE){
       refName <- gsub("_chunk_.*", "", refName)
     }
-    gffTSV <- gffTSV[which(gffTSV[, 1] == refName), ]
+      gffTSV <- gffTSV[which(gffTSV[, 1] == refName), ]
+      startPos <- elevGapSummList[[i]][[4]] * windowSize
+      endPos <- elevGapSummList[[i]][[5]] * windowSize
+    } else {
+      chunkNumber <- as.numeric(str_extract(trueRefName, "(?<=\\_)\\d+$")) - 1
+      startPos <- (elevGapSummList[[i]][[4]] * windowSize) + (chunkNumber * chunkSize)
+      endPos <- (elevGapSummList[[i]][[5]] * windowSize) + (chunkNumber * chunkSize)
     }
-    startPos <- elevGapSummList[[i]][[4]] * windowSize
-    endPos <- elevGapSummList[[i]][[5]] * windowSize
     matchRegion <- seq(startPos, endPos, 1)
     GPs <- gffTSV[which(gffTSV[, 5] %in% matchRegion), ]
     GPs$seqid <- rep(trueRefName, nrow(GPs))