+++ NEWS.md
@@ -1,3 +1,12 @@
-# ProActive 0.0.1
+## Changes in 0.1.2
++ Changed name of main function from `ProActive()` to `ProActiveDetect()`
++ Added functionality for searching read coverage plots for gene annotations of 
+interest with the geneAnnotationSearch() function
++ Fixed bug with pattern-match start and stop positions on genome/contig chunks
 
-* Initial CRAN submission.
+## Changes in 0.0.2
++ Changes for CRAN re-submission after review
+
+## Changes in 0.0.1
+
++ Initial CRAN submission.